# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nature API
  version: 0.1.0
  description: |
    Nature 백엔드 API.

    ## 이 문서의 위치

    이 파일은 `//go:embed docs/openapi.yaml` 로 바이너리에 박혀 `GET /openapi.yaml` 로
    그대로 나가고, `GET /docs` 의 Swagger UI 가 이것을 읽는다. **파일이 없으면 빌드가 깨진다.**

    사용자 인증은 초대 전용 이메일·비밀번호 방식이다. 어드민 경로는 스텁이며, 요청·응답 스키마의
    정본은 `internal/admin` 의 핸들러다 — 여기에 상세를 지어 넣으면 구현과 어긋난다.
servers:
  - url: https://nature-api.redhead.kr
    description: 운영
  - url: http://localhost:8091
    description: 로컬 개발 (Firestore 에뮬레이터 연동)
security:
  - bearerAuth: []
paths:
...
  /sms/campaigns:
    post:
      tags:
        - sms
      operationId: createSmsCampaign
      summary: SMS 캠페인 snapshot 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSmsCampaign"
      responses:
        "200":
          description: 동일 requestId 재요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
        "201":
          description: SMS 캠페인 snapshot 생성
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      description: requestId를 최초 생성 후 보존한다. 동일 입력 재요청은200 같은 캠페인, 다른 입력은409. 수신자1~50명.
    get:
      tags:
        - sms
      operationId: listSmsCampaigns
      summary: 캠페인 이력
      responses:
        "200":
          description: 캠페인 이력
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaignPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
  /sms/campaigns/{campaignId}:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - sms
      operationId: getSmsCampaign
      summary: 캠페인 진행상태
      responses:
        "200":
          description: 캠페인 진행상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /sms/campaigns/{campaignId}/recipients:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - sms
      operationId: listSmsCampaignRecipients
      summary: 발송 대상 snapshot 및 상태
      responses:
        "200":
          description: 발송 대상 snapshot 및 상태
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                properties:
                  items:
                    type: array
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/SmsCampaignRecipient"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /sms/campaigns/{campaignId}/start:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - sms
      operationId: startSmsCampaign
      summary: 명시적 발송 시작/계속 보내기
      responses:
        "200":
          description: 명시적 발송 시작/계속 보내기
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      description: READY 또는 CANCELLED 캠페인을 SENDING으로 변경. 기존 SENDING이면 멱등 응답. READY 대상이
        없으면409. 앱 기동시 자동 호출하지 않는다.
  /sms/campaigns/{campaignId}/cancel:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - sms
      operationId: cancelSmsCampaign
      summary: 후속 발송 중지
      responses:
        "200":
          description: 후속 발송 중지
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      description: 이후 claim 차단. 이미 SENDING인 한 건의 결과는 계속 저장한다. SENT는 바꾸지 않는다.
  /sms/campaigns/{campaignId}/recipients/{id}:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags:
        - sms
      operationId: updateSmsCampaignRecipient
      summary: 한 건 발송 확보 또는 결과 기록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsRecipientUpdate"
      responses:
        "200":
          description: 한 건 발송 확보 또는 결과 기록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsDispatchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      description: SENDING 요청 최초 성공만 dispatchAllowed=true. 같은 attempt 재요청은false이며
        Native 재호출 금지. 캠페인별 동시 SENDING 최대1. SENT/FAILED 결과는 같은 attempt 최종 상태 반복만
        멱등. FAILED+OUTCOME_UNKNOWN인 현재 동일 attempt의 늦은 실제 SENT 정정만 예외 허용하며 재발송 권한은
        주지 않는다. 서버 저장 성공 후 다음 대상 발송.
  /sms/campaigns/{campaignId}/recipients/{id}/retry:
    parameters:
      - name: campaignId
        in: path
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags:
        - sms
      operationId: retrySmsCampaignRecipient
      summary: 명시적으로 선택한 실패 건 재시도 준비
      responses:
        "200":
          description: 명시적으로 선택한 실패 건 재시도 준비
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsDispatchResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: ACCOUNT_DISABLED 또는 PASSWORD_CHANGE_REQUIRED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "NOT_FOUND: 내 데이터가 아니거나 없음"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: STATE_CONFLICT, PHONE_ALREADY_EXISTS 또는 IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
      description: FAILED만 READY로 변경한다. OUTCOME_UNKNOWN/PARTIAL_SENT는409로 재시도 금지.
        SENT/SENDING 거부. 최대20회, 이전 attemptId 재사용 금지. 이후 start와 새 attempt claim 필요.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        액세스 토큰(Bearer). 사용자 토큰과 어드민 토큰은 **서로 다른 시크릿으로 서명한다** —
        `ADMIN_JWT_SECRET` 이 `JWT_SECRET` 과 같으면 어드민 토큰이 사용자 미들웨어에서도
        파싱되므로 서버가 기동을 거부한다.
  responses:
...
    BadRequest:
      description: 요청 형식 또는 값 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: VALIDATION_FAILED, message: 요청 값이 올바르지 않아요 }
    Unauthorized:
      description: 토큰이 없거나 유효하지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: UNAUTHORIZED, message: 로그인이 필요해요 }
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: INTERNAL_ERROR, message: 서버 오류가 생겼어요 }
  schemas:
...
    MessageAttachment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
          enum:
          - image/jpeg
          - image/png
        size:
          type: integer
          maximum: 307200
        createdAt:
          type: string
          format: date-time
...
    CreateSmsCampaign:
      type: object
      required:
        - requestId
        - title
        - message
        - recipientIds
      properties:
        attachmentIds:
          type: array
          maxItems: 3
          uniqueItems: true
          items:
            type: string
        requestId:
          type: string
          minLength: 8
          maxLength: 128
          pattern: ^[A-Za-z0-9_-]+$
        title:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 0
          maxLength: 2000
          description: 첨부가 없으면 공백뿐인 본문 거부. 앞뒤 공백과 개행은 보존.
        recipientIds:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items:
            type: string
    SmsCampaign:
      type: object
      required:
        - id
        - title
        - message
        - status
        - recipientCount
        - readyCount
        - sendingCount
        - sentCount
        - failedCount
        - createdAt
        - updatedAt
        - startedAt
        - completedAt
      properties:
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/MessageAttachment'
        id:
          type: string
        title:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 2000
        status:
          type: string
          enum:
            - READY
            - SENDING
            - COMPLETED
            - PARTIAL_FAILED
            - CANCELLED
        recipientCount:
          type: integer
          minimum: 0
        readyCount:
          type: integer
          minimum: 0
        sendingCount:
          type: integer
          minimum: 0
        sentCount:
          type: integer
          minimum: 0
        failedCount:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    SmsCampaignPage:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SmsCampaign"
        nextCursor:
          type: string
          nullable: true
    SmsCampaignRecipient:
      type: object
      required:
        - id
        - campaignId
        - recipientId
        - name
        - phone
        - message
        - status
        - attemptId
        - attemptCount
        - sentAt
        - failedAt
        - errorCode
        - errorMessage
        - createdAt
        - updatedAt
      properties:
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/MessageAttachment'
        transport:
          type: string
          enum:
          - SMS
          - LMS
          - MMS
        id:
          type: string
        campaignId:
          type: string
        recipientId:
          type: string
        name:
          type: string
          maxLength: 100
        phone:
          type: string
        message:
          type: string
          maxLength: 2000
        status:
          type: string
          enum:
            - READY
            - SENDING
            - SENT
            - FAILED
        attemptId:
          type: string
        attemptCount:
          type: integer
          minimum: 0
        sentAt:
          type: string
          format: date-time
          nullable: true
        failedAt:
          type: string
          format: date-time
          nullable: true
        errorCode:
          type: string
          maxLength: 100
        errorMessage:
          type: string
          maxLength: 300
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SmsRecipientUpdate:
      type: object
      required:
        - status
        - attemptId
      properties:
        transport:
          type: string
          enum:
          - SMS
          - LMS
          - MMS
        status:
          type: string
          enum:
            - SENDING
            - SENT
            - FAILED
        attemptId:
          type: string
          minLength: 8
          maxLength: 128
          pattern: ^[A-Za-z0-9_-]+$
        errorCode:
          type: string
          maxLength: 100
          description: FAILED 필수. OUTCOME_UNKNOWN/PARTIAL_SENT는 확인필요 및 재시도 금지.
        errorMessage:
          type: string
          maxLength: 300
    SmsDispatchResult:
      type: object
      required:
        - campaign
        - recipient
        - dispatchAllowed
      properties:
        campaign:
          $ref: "#/components/schemas/SmsCampaign"
        recipient:
          $ref: "#/components/schemas/SmsCampaignRecipient"
        dispatchAllowed:
          type: boolean
          description: 최초 READY→SENDING 트랜잭션 승자만 true. false면 SMS를 호출하지 않는다.
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계 판독용 에러 코드. 클라이언트는 이 값으로 분기하고 `message` 는 그대로 표시한다.
            공통 코드는 `VALIDATION_FAILED` · `UNAUTHORIZED` · `INTERNAL_ERROR` 이며,
            사용자 인증은 `INVALID_CREDENTIALS` · `ACCOUNT_DISABLED`를 더한다.
            `TOO_MANY_ATTEMPTS`는 향후 시도 제한용 429 예약 코드다. 현재 제한은 구현하지 않았으며,
            구현 시 `details`에 남은 초를 담는다.
          example: VALIDATION_FAILED
        message:
          type: string
          description: 사용자에게 그대로 보여줄 수 있는 한국어 문구
          example: 요청 값이 올바르지 않아요
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: 필드별 위반 내역 등 부가 정보
...